refactor(proxy): pass APIType to pdConnectorRunner and deduplicate tests
- Change pdConnectorRunner signature from []string to APIType so each
  connector decides internally which JSON fields it needs, keeping the
  abstraction boundary clean.
- NIXL v2 resolves token limit fields internally via
  tokenLimitFieldsForAPIType; SharedStorage and SGLang ignore APIType.
- Replace per-request map allocation in NIXL v2 with a zero-alloc
  fixed-size array for saving/restoring token limit fields.
- Consolidate duplicated prefill-header routing tests into a single
  testPrefillHeaderRouting helper parameterized by APIType.
- Merge connector_responses_test.go and responses_test.go into
  connector_nixlv2_test.go and handler_test.go respectively.
